Added test for verify that user can click the sign in button; Added headless mode as a ChromeOption
