+++ Test/packages.config
@@ -4,6 +4,7 @@
   <package id="Microsoft.Win32.Registry" version="5.0.0" targetFramework="net472" />
   <package id="NUnit" version="3.12.0" targetFramework="net472" />
   <package id="NUnit3TestAdapter" version="3.17.0" targetFramework="net472" />
+  <package id="Selenium.Support" version="3.141.0" targetFramework="net472" />
   <package id="Selenium.WebDriver" version="3.141.0" targetFramework="net472" />
   <package id="Selenium.WebDriver.ChromeDriver" version="87.0.4280.8800" targetFramework="net472" />
   <package id="SharpZipLib" version="1.3.1" targetFramework="net472" />
