+++ greencitytests.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30804.86
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SeleniumEasyTest", "Test\SeleniumEasyTest.csproj", "{7AC19374-81D2-4A24-97A4-57AEE9F984C9}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GreenCityTests", "Test\GreenCityTests.csproj", "{7AC19374-81D2-4A24-97A4-57AEE9F984C9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
